feat: HIGH security fixes - DEBUG toggle, rate limiting, DecimalField, auth on ViewSets
- DEBUG now controlled by DJANGO_DEBUG env var (default: false)
- DRF throttling: 60 req/min for anonymous users
- All monetary fields changed from FloatField to DecimalField (24 fields across 4 models)
- Migration 0007 for FloatField->DecimalField conversion
- RequireAuthForDestructive permission: GET/POST open, PUT/PATCH/DELETE require Bearer token
- CalculationViewSet, DailyLogViewSet, WalletViewSet, TransactionViewSet all protected
- Frontend: unified getAuthHeaders() helper, all DELETE/PUT calls include auth token
- Bumped version to 1.5.0

+++ backend/calculator/models.py
@@ -3,33 +3,33 @@ from django.db import models
 
 class Calculation(models.Model):
     label = models.CharField(max_length=100)
-    capital = models.FloatField()
-    comision = models.FloatField(default=0.35)  # Legacy field
+    capital = models.DecimalField(max_digits=14, decimal_places=2)
+    comision = models.DecimalField(max_digits=6, decimal_places=4, default=0.35)  # Legacy field
     
     # Multi-Currency / Multi-Platform fields
     tipo_operativa = models.CharField(max_length=20, default='USD')
     plataforma_compra = models.CharField(max_length=50, default='Binance P2P')
     plataforma_venta = models.CharField(max_length=50, default='Binance P2P')
-    comision_compra = models.FloatField(default=0.35)
-    comision_venta = models.FloatField(default=0.35)
+    comision_compra = models.DecimalField(max_digits=6, decimal_places=4, default=0.35)
+    comision_venta = models.DecimalField(max_digits=6, decimal_places=4, default=0.35)
     metodo_compra = models.CharField(max_length=50, default='Zinli')
     metodo_venta = models.CharField(max_length=50, default='Zinli')
     
-    tasa_venta = models.FloatField()
-    tasa_compra = models.FloatField()
-    tasa_retorno = models.FloatField(default=1.0)
+    tasa_venta = models.DecimalField(max_digits=12, decimal_places=4)
+    tasa_compra = models.DecimalField(max_digits=12, decimal_places=4)
+    tasa_retorno = models.DecimalField(max_digits=6, decimal_places=4, default=1.0)
     
     ciclos_dia = models.IntegerField(default=1)
     metodos_pago = models.IntegerField(default=1)
     
     # Pre-calculated values
-    monto_venta = models.FloatField()
-    monto_compra = models.FloatField()
-    ganancia_porcentaje = models.FloatField()
-    ganancia_ciclo = models.FloatField()
-    ganancia_diaria = models.FloatField()
-    ganancia_mensual = models.FloatField()
-    tasa_minima_compra = models.FloatField()
+    monto_venta = models.DecimalField(max_digits=14, decimal_places=2)
+    monto_compra = models.DecimalField(max_digits=14, decimal_places=2)
+    ganancia_porcentaje = models.DecimalField(max_digits=8, decimal_places=4)
+    ganancia_ciclo = models.DecimalField(max_digits=14, decimal_places=2)
+    ganancia_diaria = models.DecimalField(max_digits=14, decimal_places=2)
+    ganancia_mensual = models.DecimalField(max_digits=14, decimal_places=2)
+    tasa_minima_compra = models.DecimalField(max_digits=12, decimal_places=4)
     
     created_at = models.DateTimeField(auto_now_add=True)
 
@@ -42,16 +42,16 @@ class Calculation(models.Model):
 
 class DailyLog(models.Model):
     date = models.DateField()
-    profit = models.FloatField(default=0.0)
-    volume = models.FloatField(default=0.0)
+    profit = models.DecimalField(max_digits=14, decimal_places=2, default=0.0)
+    volume = models.DecimalField(max_digits=14, decimal_places=2, default=0.0)
     imported = models.BooleanField(default=False)
     notes = models.TextField(blank=True, default='')
     
     tipo_operativa = models.CharField(max_length=20, default='USD')
     plataforma_compra = models.CharField(max_length=50, default='Binance P2P')
     plataforma_venta = models.CharField(max_length=50, default='Binance P2P')
-    comision_compra = models.FloatField(default=0.35)
-    comision_venta = models.FloatField(default=0.35)
+    comision_compra = models.DecimalField(max_digits=6, decimal_places=4, default=0.35)
+    comision_venta = models.DecimalField(max_digits=6, decimal_places=4, default=0.35)
     metodo_compra = models.CharField(max_length=50, default='Zinli')
     metodo_venta = models.CharField(max_length=50, default='Zinli')
 
@@ -72,8 +72,8 @@ class Wallet(models.Model):
     name = models.CharField(max_length=100)  # e.g. "Binance USDT Principal"
     platform = models.CharField(max_length=50)  # e.g. "Binance", "Zinli", "Banesco"
     currency = models.CharField(max_length=10, choices=CURRENCY_CHOICES)
-    balance = models.FloatField(default=0.0)
-    opening_balance = models.FloatField(default=0.0)
+    balance = models.DecimalField(max_digits=14, decimal_places=2, default=0.0)
+    opening_balance = models.DecimalField(max_digits=14, decimal_places=2, default=0.0)
     is_active = models.BooleanField(default=True)
     color = models.CharField(max_length=7, default='#2563eb')  # Hex color for UI
     created_at = models.DateTimeField(auto_now_add=True)
@@ -112,11 +112,11 @@ class Transaction(models.Model):
         related_name='transactions_in'
     )
     
-    amount_out = models.FloatField(default=0.0)  # Sale de wallet_from
-    amount_in = models.FloatField(default=0.0)   # Entra a wallet_to
+    amount_out = models.DecimalField(max_digits=14, decimal_places=2, default=0.0)  # Sale de wallet_from
+    amount_in = models.DecimalField(max_digits=14, decimal_places=2, default=0.0)   # Entra a wallet_to
     
-    rate = models.FloatField(default=1.0)  # Tasa de cambio utilizada
-    commission_pct = models.FloatField(default=0.0)  # Comisión P2P %
+    rate = models.DecimalField(max_digits=12, decimal_places=4, default=1.0)  # Tasa de cambio utilizada
+    commission_pct = models.DecimalField(max_digits=6, decimal_places=4, default=0.0)  # Comisión P2P %
     
     notes = models.TextField(blank=True, default='')
     created_at = models.DateTimeField(auto_now_add=True)
